+++ addons/yat/src/interfaces/IDebugScreenItem.cs
@@ -1,0 +1,6 @@
+namespace YAT.Interfaces;
+
+public interface IDebugScreenItem
+{
+	public void Update();
+}