+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 -   DebugScreen scene.
 -   EDebugScreenItemPosition enum.
+-   IDebugScreenItem interface.
 
 ## [1.24.1-beta 2024-02-24]